refactor(extension): fix prod build with aot & barel problem
