+++ src/app/modules/angular-slickgrid/services/gridState.service.ts
@@ -10,7 +10,9 @@ import {
   GridStateChange,
   GridStateType,
 } from './../models/index';
-import { ExtensionService, FilterService, SortService } from './../services/index';
+import { ExtensionService } from './extension.service';
+import { FilterService } from './filter.service';
+import { SortService } from './sort.service';
 import { Subject } from 'rxjs/Subject';
 import { Subscription } from 'rxjs/Subscription';
